Removes unnecessary 'await' before 'return'
Dart, as of version 2.1.1, has a linter rule called
'unnecessary_await_in_return', described as 'Avoid returning
an awaited expression when the expression type is assignable
to the function's return type'. This can be found at
https://dart.dev/tools/linter-rules/unnecessary_await_in_return
VSCode applied that linter rule automatically, and I am tired of
the warnings, so I'm submitting this. :)